ai-learning-backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astunparse==1.6.3
bcrypt==3.2.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cloudinary==1.44.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dnspython==2.7.0
python-dotenv==1.1.1
ecdsa==0.19.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.5
gast==0.6.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.37.1
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.12.0
keras==3.10.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-groq==0.3.7
langchain-openai==0.3.28
langgraph==0.6.4
langsmith==0.4.13
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
motor==3.7.1
namex==0.1.0
numpy==2.1.3
openai==1.99.9
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
passlib[bcrypt]==1.7.4
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
PyNaCl==1.5.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn[standard]==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.6.0
zstandard==0.23.0